Refactor tmdb_exceptions -> exceptions
Refactor tmdb -> api
