+++ medusa/indexers/indexer_config.py
@@ -3,7 +3,7 @@
 """Indexer config module."""
 
 from medusa.app import BASE_PYMEDUSA_URL
-from medusa.indexers.tmdb.tmdb import Tmdb
+from medusa.indexers.tmdb.api import Tmdb
 from medusa.indexers.tvdb.api import TVDBv2
 from medusa.indexers.tvmaze.tvmaze_api import TVmaze
 from medusa.session.core import MedusaSession
